+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 - CI-498 - Changed ebook pdf to a cta link
 - XOT-1039 - Add campaign block to Great Domestic Homepage
 - No ticket - Add database fixtures and update readme with usage instructions
-
+- CI-503 - Added fields to WhyChooseUkPage for ready to trade
 
 ## [2019.09.18](https://github.com/uktrade/directory-cms/releases/tag/2019.09.18)
 [Full Changelog](https://github.com/uktrade/directory-cms/compare/2019.08.27_1...2019.09.18)